excel-mcp-backend/requirements.txt
mcp[cli]>=1.2.0
xlwings>=0.30.0
openpyxl>=3.1.2
fastapi>=0.95.0
uvicorn>=0.21.0
# Ne pas inclure torch, il sera installé dans le Dockerfile
python-multipart>=0.0.6
pyyaml>=6.0
requests>=2.28.2
#transformers>=4.28.1
sentence-transformers>=2.2.2
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0